Upload V0.1
Developer to Manager

Interviews with Software Managers About transitioning from development to management

Link : https://devtomanager.com/interviews/

Procedure:

To extract data using web scraping with python, you need to follow these basic steps:

--->Find the URL that you want to scrape --->Inspecting the Page --->Find the data you want to extract --->Write the code --->Run the code and extract the data --->Store the data in the required format

Libraries used for Web Scraping

As we know, Python is has various applications and there are different libraries for different purposes. In our further demonstration, we will be using the following libraries:

Selenium: Selenium is a web testing library. It is used to automate browser activities.

BeautifulSoup: Beautiful Soup is a Python package for parsing HTML and XML documents. It creates parse trees that is helpful to extract the data easily.

Pandas: Pandas is a library used for data manipulation and analysis. It is used to extract the data and store it in the desired format.